分离类型定义到多个文件中
